+++ tee-worker/client-api/parachain-api/prepare-build/interfaces/definitions.ts
@@ -1,4 +1,4 @@
 export { default as identity } from "./identity/definitions";
 export { default as vc } from "./vc/definitions";
-export { default as trustedCalls } from "./trusted-calls/definitions";
+export { default as trusted_calls } from "./trusted_calls/definitions";
 export { default as sidechain } from "./sidechain/definitions";
