+++ tee-worker/client-api/parachain-api/src/index.ts
@@ -14,6 +14,6 @@ export * from "@polkadot/api";
 export * from "@polkadot/api/types";
 import { default as identity } from "../build/interfaces/identity/definitions";
 import { default as vc } from "../build/interfaces/vc/definitions";
-import { default as trustedCalls } from "../build/interfaces/trusted-calls/definitions";
+import { default as trusted_calls } from "../build/interfaces/trusted_calls/definitions";
 import { default as sidechain } from "../build/interfaces/sidechain/definitions";
-export { identity, vc, trustedCalls, sidechain };
+export { identity, vc, trusted_calls, sidechain };
